# FileTarget - Faster archive check by caching the static file-create-time (60-70% improvement)