FileTarget - FileAppenderCache now caches the conversion of File-Time-Utc to File-Time-TimeSource thus avoiding the expensive DateTime.ToLocalTime(). Also the ArchiveEvery-check now performs truncate of DateTime.Ticks instead of performing DateTime.ToString() allocation.

This PR also contains the following changes:
- MutexMultiProcessFileAppender now uses the cached File-Creation-Time, instead of doing lookup using file-handle
- WindowsMultiProcessFileAppender now uses the cached File-Creation-Time, instead of doing lookup using file-handle
- UnixMultiProcessFileAppender now updates the cached File-Creation-Time as expected by the FileTarget / FileAppenderCache.
- Changed some FileTargetTests to use ShiftedTimeSource instead of using File.SetCreationTime. This allows FileAppenders that caches the static File-Creation-Time to be tested (Now also testing SingleProcessFileAppender).

Have done performance tests with the following config:

`<target name="SyncFilePerfArchive" type="File" fileName="C:\Temp\Log\NLogPerformanceArchivePerf.txt" autoFlush="false" keepFileOpen="true" concurrentWrites="false" layout="${message}" ArchiveAboveSize="10000000000" MaxArchiveFiles="1000" ArchiveEvery="Year" ArchiveNumbering="Date" />`

| Logger Name      | Messages   | Size | Threads | Loggers |
|------------------|------------|------|---------|---------|
| PerfArchiveLogger | 10.000.000 |  128 |       1 |       1 |

| This PR | Time (ms) | Msgs/sec  | GC2 | GC1 | GC0 | CPU (ms) | Mem (MB) |
|------------|-----------|-----------|-----|-----|-----|----------|----------|
| concurrentWrites = false   |    18.694 |   534.917 |   0 |   0 | 1243 |   19.000 |   36,551 |
| concurrentWrites = true    |    90.754 |   110.187 |   0 |   0 | 1275 |   90.843 |   36,320 |

| Master | Time (ms) | Msgs/sec  | GC2 | GC1 | GC0 | CPU (ms) | Mem (MB) |
|------------|-----------|-----------|-----|-----|-----|----------|----------|
| concurrentWrites = false   |    32.902 |   303.924 |   0 |   1 | 1459 |   33.046 |   36,746 |
| concurrentWrites = true    |    144.539 |    69.185 |   0 |   1 | 1520 |  143.015 |   36,648 |

Merry Kerstmis :)

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/nlog/nlog/1876)
<!-- Reviewable:end -->