@304NotModified 2017-01-03
Reviewed 10 of 10 files at r1.
Review status: all files reviewed at latest revision, 9 unresolved discussions.

---

*[src/NLog/Internal/FileAppenders/BaseFileAppender.cs, line 79 at r1](https://reviewable.io:443/reviews/nlog/nlog/1876#-K_aLtP_ufvVSnCMswR0:-K_aLtP_ufvVSnCMswR1:bk5kusc) ([raw file](https://github.com/nlog/nlog/blob/5c8cbe2ff6818f490e49179ac82351512a067b32/src/NLog/Internal/FileAppenders/BaseFileAppender.cs#L79)):*
> ```Smalltalk
>         /// </summary>
>         /// <returns>The creation time of the file.</returns>
>         public DateTime CreationTimeUtc { get { return _creationTimeUtc; } internal set { _creationTimeUtc = value; CreationTimeSouce = Time.TimeSource.Current.FromSystemTime(value); } }
> ```

smart one. Please add a comment that we are doining this for performance reasons.

---

*[src/NLog/Internal/FileAppenders/BaseFileAppender.cs, line 82 at r1](https://reviewable.io:443/reviews/nlog/nlog/1876#-K_aLy7nJ6vyPFmOEoWG:-K_aLy7nJ6vyPFmOEoWH:b-7prn9v) ([raw file](https://github.com/nlog/nlog/blob/5c8cbe2ff6818f490e49179ac82351512a067b32/src/NLog/Internal/FileAppenders/BaseFileAppender.cs#L82)):*
> ```Smalltalk
>         DateTime _creationTimeUtc;
> 
>         public DateTime CreationTimeSouce { get; private set; }
> ```

Please add some XML docs and refer to CreationTimeUtc (<see tag)

It's also SouRce? (or is it my limited English vocabulary..)

---

*[src/NLog/Internal/FileAppenders/FileAppenderCache.cs, line 177 at r1](https://reviewable.io:443/reviews/nlog/nlog/1876#-K_aLm89klMqJntk6ZYH:-K_aLm89klMqJntk6ZYI:b-i9qdya) ([raw file](https://github.com/nlog/nlog/blob/5c8cbe2ff6818f490e49179ac82351512a067b32/src/NLog/Internal/FileAppenders/FileAppenderCache.cs#L177)):*
> ```Smalltalk
>                 }
> 
>                 if (string.CompareOrdinal(appenders[i].FileName, fileName) == 0)
> ```

What about .Compare (...Ordinal)? 

---

*[src/NLog/Internal/FileAppenders/FileAppenderCache.cs, line 182 at r1](https://reviewable.io:443/reviews/nlog/nlog/1876#-K_aMUE8RGEVj-woHq_q:-K_aMUE8RGEVj-woHq_r:bgaoe8t) ([raw file](https://github.com/nlog/nlog/blob/5c8cbe2ff6818f490e49179ac82351512a067b32/src/NLog/Internal/FileAppenders/FileAppenderCache.cs#L182)):*
> ```Smalltalk
>                     // (MRU)
>                     BaseFileAppender app = appenders[i];
>                     if (i != 0)
> ```

if (i > 0)? 

---

*[src/NLog/Internal/FileAppenders/FileAppenderCache.cs, line 332 at r1](https://reviewable.io:443/reviews/nlog/nlog/1876#-K_aMp_Rp7RAp5x3VfqN:-K_aMp_Rp7RAp5x3VfqO:b9sfubg) ([raw file](https://github.com/nlog/nlog/blob/5c8cbe2ff6818f490e49179ac82351512a067b32/src/NLog/Internal/FileAppenders/FileAppenderCache.cs#L332)):*
> ```Smalltalk
>                     if (result.HasValue)
>                     {
>                         if (result.Value != appender.CreationTimeUtc)
> ```

we're checking the cached version here right? Please add it as comment, thanks!

---

*[src/NLog/Internal/FileAppenders/MutexMultiProcessFileAppender.cs, line 196 at r1](https://reviewable.io:443/reviews/nlog/nlog/1876#-K_aMxkbLvpLoYcW6p4P:-K_aMxkbLvpLoYcW6p4Q:b-tf1cn7) ([raw file](https://github.com/nlog/nlog/blob/5c8cbe2ff6818f490e49179ac82351512a067b32/src/NLog/Internal/FileAppenders/MutexMultiProcessFileAppender.cs#L196)):*
> ```Smalltalk
>         public override DateTime? GetFileCreationTimeUtc()
>         {
>             return CreationTimeUtc; // File is kept open, so creation time is static
> ```

smart one :)

Thanks for the comment. Needed it

---

*[src/NLog/Internal/FileAppenders/UnixMultiProcessFileAppender.cs, line 101 at r1](https://reviewable.io:443/reviews/nlog/nlog/1876#-K_aNEvbdlv6Lf2eX2Us:-K_aNEvbdlv6Lf2eX2Ut:b-zh1gu5) ([raw file](https://github.com/nlog/nlog/blob/5c8cbe2ff6818f490e49179ac82351512a067b32/src/NLog/Internal/FileAppenders/UnixMultiProcessFileAppender.cs#L101)):*
> ```Smalltalk
> 
>                 long filePosition = this.file.Position;
>                 if (fileExists || filePosition > 0)
> ```

isn't fileExists always true when filePosition > 0 ?

---

*[src/NLog/Internal/FileAppenders/WindowsMultiProcessFileAppender.cs, line 203 at r1](https://reviewable.io:443/reviews/nlog/nlog/1876#-K_aNXBfxYakFafM0Bhv:-K_aNXBfxYakFafM0Bhw:b-qgnq3p) ([raw file](https://github.com/nlog/nlog/blob/5c8cbe2ff6818f490e49179ac82351512a067b32/src/NLog/Internal/FileAppenders/WindowsMultiProcessFileAppender.cs#L203)):*
> ```Smalltalk
>         public override DateTime? GetFileCreationTimeUtc()
>         {
>             return CreationTimeUtc; // File is kept open, so creation time is static
> ```

are you sure this is the case here?

---

*[src/NLog/Targets/FileTarget.cs, line 2065 at r1](https://reviewable.io:443/reviews/nlog/nlog/1876#-K_aNfRVLKq0AyUAtLWP:-K_aNfRVLKq0AyUAtLWQ:b-xtpyir) ([raw file](https://github.com/nlog/nlog/blob/5c8cbe2ff6818f490e49179ac82351512a067b32/src/NLog/Targets/FileTarget.cs#L2065)):*
> ```Smalltalk
>         }
> 
>         private static DateTime TruncateArchiveTime(DateTime input, FileArchivePeriod resolution)
> ```

what are we trying to do here? (xmldoc please)

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/nlog/nlog/1876)*
<!-- Sent from Reviewable.io -->