@snakefoot 2017-01-03
*[src/NLog/Internal/FileAppenders/FileAppenderCache.cs, line 182 at r1](https://reviewable.io:443/reviews/nlog/nlog/1876#-K_aMUE8RGEVj-woHq_q:-K_aWruUTwmGvQoibT3Z:b63mq4u) ([raw file](https://github.com/nlog/nlog/blob/5c8cbe2ff6818f490e49179ac82351512a067b32/src/NLog/Internal/FileAppenders/FileAppenderCache.cs#L182)):*
<details><summary><i>Previously, 304NotModified (Julian Verdurmen) wrote…</i></summary><blockquote>

if (i > 0)? 
</blockquote></details>

Also works

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/nlog/nlog/1876)*
<!-- Sent from Reviewable.io -->