@snakefoot 2017-01-03
*[src/NLog/Internal/FileAppenders/WindowsMultiProcessFileAppender.cs, line 203 at r1](https://reviewable.io:443/reviews/nlog/nlog/1876#-K_aNXBfxYakFafM0Bhv:-K_aY-dXeSXTn6567Xc6:bm5sq27) ([raw file](https://github.com/nlog/nlog/blob/5c8cbe2ff6818f490e49179ac82351512a067b32/src/NLog/Internal/FileAppenders/WindowsMultiProcessFileAppender.cs#L203)):*
<details><summary><i>Previously, 304NotModified (Julian Verdurmen) wrote…</i></summary><blockquote>

are you sure this is the case here?
</blockquote></details>

FileCreation-time should not be updated by having multiple writers to the same file. They will just update FileModified-time. Ofcourse it will not longer react to random application that for fun modifies the file-creation-time of log-files.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/nlog/nlog/1876)*
<!-- Sent from Reviewable.io -->