@snakefoot 2017-01-03
*[src/NLog/Internal/FileAppenders/UnixMultiProcessFileAppender.cs, line 101 at r1](https://reviewable.io:443/reviews/nlog/nlog/1876#-K_aNEvbdlv6Lf2eX2Us:-K_aZrDSN_1dUDCoSzrs:bo31sib) ([raw file](https://github.com/nlog/nlog/blob/5c8cbe2ff6818f490e49179ac82351512a067b32/src/NLog/Internal/FileAppenders/UnixMultiProcessFileAppender.cs#L101)):*
<details><summary><i>Previously, 304NotModified (Julian Verdurmen) wrote…</i></summary><blockquote>

isn't fileExists always true when filePosition > 0 ?
</blockquote></details>

Well there can be a race-condition between the fileExists-check and the filePosition > 0. If someone was able to create the file and append to it between doing the fileExists-check and the filePosition-check.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/nlog/nlog/1876)*
<!-- Sent from Reviewable.io -->